+++ tests/integration/Contrib.License.Factories.LicenseFactory.test.js
@@ -20,15 +20,18 @@ Copyright (c) OWASP Foundation. All Rights Reserved.
 const assert = require('node:assert')
 
 const { suite, test } = require('mocha')
+const spdxExpressionValidate = require('spdx-expression-parse')
 
 const {
-  Factories: { LicenseFactory },
+  Contrib,
   Models: { LicenseExpression, NamedLicense, SpdxLicense }
 } = require('../../')
 
-suite('integration: Factories.LicenseFactory', () => {
+suite('integration: Contrib.License.Factories.LicenseFactory', () => {
   test('makeFromString() -> LicenseExpression', () => {
-    const sut = new LicenseFactory()
+    const sut = new Contrib.License.Factories.LicenseFactory(
+      spdxExpressionValidate
+    )
     const expression = '(MIT OR Apache-2.0)'
 
     const license = sut.makeFromString(expression)
@@ -38,7 +41,7 @@ suite('integration: Factories.LicenseFactory', () => {
   })
 
   test('makeFromString() -> NamedLicense', () => {
-    const sut = new LicenseFactory()
+    const sut = new Contrib.License.Factories.LicenseFactory()
 
     const license = sut.makeFromString('(c) foo bar')
 
@@ -49,7 +52,7 @@ suite('integration: Factories.LicenseFactory', () => {
   })
 
   test('makeFromString() -> SpdxLicense', () => {
-    const sut = new LicenseFactory()
+    const sut = new Contrib.License.Factories.LicenseFactory()
 
     const license = sut.makeFromString('MIT')